Unify into a single header-only library; drop the assembly
The bench showed the pure-C++ hot path matches the assembly, so retire the
asm entirely and collapse everything into one header-only library.

include/profiler.h is now the whole library (stb-style):
- Light section (always compiled): Event/EventBuffer/CustomTLS, the inline
  emit_* hot path (faithful port of the old asm: segment-tid -> custom_tls[]
  index, single rdtsc with the +1/+5/+10 slot offsets, double-buffer bounds
  check), and the config macros - now the single source of truth, overridable
  via -D / #define.
- Implementation section behind #ifdef LOP_IMPLEMENTATION (one TU): the engine
  (threads, flush, spill, windows.h). enabled/custom_tls became lightweight
  inline globals the hot path reads; everything else stays engine-side.

Deleted src/profiler.cpp, src/profiler_asm.cpp, src/profiler_asm.asm and
src/profiler_cpp.cpp (the src/ tree is gone). Samples and tests each define
LOP_IMPLEMENTATION (every standalone program is its own one TU).

Tooling: run_tests.sh / run_bench.sh and CI now build header-only with config
via -D (no source patching, no ml64/MASM); CI gains a multi-TU ODR check and
the Windows job is pure MSVC. README/test docs updated for header-only usage.

Verified: gcc+clang -Wall -Wextra clean, samples produce valid traces,
multi-TU ODR links, the full spill suite passes (incl. the OOM relief valve),
and per-event overhead is unchanged (~24 ns local).
