+++ test/run_bench.sh
@@ -1,11 +1,10 @@
 #!/usr/bin/env bash
 #
-# Compare per-event tracing overhead of the assembly backend vs the pure-C++ backend.
+# Measure per-event tracing overhead of the (header-only) profiler.
 #
-# Builds the SAME test/bench.cpp against each backend, runs both, and prints a comparison of
-# per-event overhead (wall-time delta of tracing on vs off, per event) and the minimal observed
-# gap between two consecutive events. Gates on correctness (both backends must emit the same,
-# balanced event count) but NOT on the perf numbers - those are advisory, especially on a VM.
+# Builds test/bench.cpp, runs it, and reports per-event overhead (wall-time delta of tracing on vs
+# off, per event) and the smallest observed gap between two consecutive events. Gates on correctness
+# (balanced begin/end counts) but NOT on the perf numbers - those are advisory, especially on a VM.
 #
 # Linux/x64; needs g++ (override with CXX=) and python3. Scale work with LOP_BENCH_PAIRS / LOP_BENCH_REPS.
 set -euo pipefail
@@ -18,54 +17,37 @@ WORK="$(mktemp -d)"
 trap 'rm -rf "$WORK"' EXIT
 fail() { echo "FAILED: $*" >&2; exit 1; }
 
-echo "Building bench (asm backend) ..."
-$CXX $CXXFLAGS "$HERE/bench.cpp" "$ROOT/src/profiler_asm.cpp" "$ROOT/src/profiler.cpp" -o "$WORK/bench_asm" || fail "asm build"
-echo "Building bench (c++ backend) ..."
-$CXX $CXXFLAGS "$HERE/bench.cpp" "$ROOT/src/profiler_cpp.cpp" "$ROOT/src/profiler.cpp" -o "$WORK/bench_cpp" || fail "cpp build"
+echo "Building bench ..."
+$CXX $CXXFLAGS "$HERE/bench.cpp" -o "$WORK/bench" || fail "build"
 
-run_backend() { # <name> <binary>
-    local name="$1"
-    local bin="$2"
-    local dir="$WORK/$name"
-    mkdir -p "$dir"
-    ( cd "$dir" && "$bin" >"$dir/run.txt" 2>/dev/null ) || fail "$name run"
-    python3 "$HERE/bench_report.py" "$dir" >>"$dir/run.txt" || fail "$name report"
-}
+( cd "$WORK" && "$WORK/bench" >"$WORK/run.txt" 2>/dev/null ) || fail "run"
+python3 "$HERE/bench_report.py" "$WORK" >>"$WORK/run.txt" || fail "report"
 
-run_backend asm "$WORK/bench_asm"
-run_backend cpp "$WORK/bench_cpp"
+kv()   { grep -oE "$1=[0-9.-]+" "$WORK/run.txt" | head -1 | cut -d= -f2; }  # BENCH line "key=val"
+line() { grep -E "^$1 " "$WORK/run.txt" | head -1 | awk '{print $2}'; }      # report line "KEY val"
 
-kv()   { grep -oE "$2=[0-9.-]+" "$1" | head -1 | cut -d= -f2; }  # BENCH line "key=val"
-line() { grep -E "^$2 " "$1" | head -1 | awk '{print $2}'; }     # report line "KEY val"
+PE=$(kv per_event_ns)
+MIN=$(line MIN_CONSECUTIVE_NS)
+MED=$(line MEDIAN_CONSECUTIVE_NS)
+EV=$(line EVENTS)
+BAL=$(line BALANCED)
 
-A="$WORK/asm/run.txt"; C="$WORK/cpp/run.txt"
-A_PE=$(kv "$A" per_event_ns);  C_PE=$(kv "$C" per_event_ns)
-A_MIN=$(line "$A" MIN_CONSECUTIVE_NS); C_MIN=$(line "$C" MIN_CONSECUTIVE_NS)
-A_MED=$(line "$A" MEDIAN_CONSECUTIVE_NS); C_MED=$(line "$C" MEDIAN_CONSECUTIVE_NS)
-A_EV=$(line "$A" EVENTS); C_EV=$(line "$C" EVENTS)
-A_BAL=$(line "$A" BALANCED); C_BAL=$(line "$C" BALANCED)
+[ "$BAL" = yes ] || fail "begin/end not balanced"
 
-# Correctness gate (perf numbers are not gated).
-[ "$A_BAL" = yes ] && [ "$C_BAL" = yes ] || fail "begin/end not balanced (asm=$A_BAL cpp=$C_BAL)"
-[ "$A_EV" = "$C_EV" ] || fail "event counts differ (asm=$A_EV cpp=$C_EV)"
-
-printf '\n%-8s %16s %16s %18s\n' backend per_event_ns min_consec_ns median_consec_ns
-printf '%-8s %16s %16s %18s\n' asm "$A_PE" "$A_MIN" "$A_MED"
-printf '%-8s %16s %16s %18s\n' cpp "$C_PE" "$C_MIN" "$C_MED"
+printf '\n%18s %16s %18s %8s\n' per_event_ns min_consec_ns median_consec_ns events
+printf '%18s %16s %18s %8s\n' "$PE" "$MIN" "$MED" "$EV"
 echo
-echo "Correctness: both balanced, $A_EV events each."
 echo "(Per-event = wall delta on/off per event; min_consec = closest two timestamps. Noisy on VMs.)"
 
 if [ -n "${GITHUB_STEP_SUMMARY:-}" ]; then
     {
-        echo "### Tracing overhead: asm vs C++ backend"
+        echo "### Tracing overhead (header-only C++)"
         echo ""
-        echo "| backend | per-event ns (Δ on/off) | min consecutive ns | median consecutive ns |"
+        echo "| per-event ns (Δ on/off) | min consecutive ns | median consecutive ns | events |"
         echo "|---|---|---|---|"
-        echo "| asm | $A_PE | $A_MIN | $A_MED |"
-        echo "| cpp | $C_PE | $C_MIN | $C_MED |"
+        echo "| $PE | $MIN | $MED | $EV |"
         echo ""
-        echo "_Events: $A_EV each. Numbers are noisy on shared CI runners; treat as indicative, not a gate._"
+        echo "_Numbers are noisy on shared CI runners; treat as indicative, not a gate._"
     } >> "$GITHUB_STEP_SUMMARY"
 fi
